+++ ReleaseNotes.md
@@ -1,4 +1,8 @@
 
+# 1.05
+
+* add option to use one note for whole team (like before)
+* add option to enable note per team-member
 
 # 1.04